added controller/things/create_test.
