+++ test/controllers/things_controller_test.rb
@@ -24,107 +24,9 @@ class ThingsControllerTest < IntegrationTest
     thing
   end
 
-  describe "#new" do
-    it "#new [HTML]" do
-      visit "/things/new"
-
-      page.must_have_css "form #thing_name"
-      page.wont_have_css "form #thing_name.readonly"
-
-      # 3 author email fields
-      page.must_have_css("input.email", count: 3) # TODO: how can i say "no value"?
-      page.wont_have_css("#thing_is_author")
-    end
-
-    # SignedIn
-    it "what" do
-      sign_in!
-
-      visit "/things/new"
-
-      page.must_have_css "form #thing_name"
-      page.wont_have_css "form #thing_name.readonly"
-
-      # 3 author email fields
-      page.must_have_css("input.email", count: 3) # TODO: how can i say "no value"?
-      page.must_have_css("#thing_is_author")
-
-      fill_in 'Name', with: "Bad Religion"
-      check "I'm the author!"
-      click_button "Create Thing"
-
-      # /things/1
-      page.current_path.must_equal thing_path(Thing.last)
-      page.must_have_content "By fred@trb.org"
-
-      click_link "Edit"
-
-      # /things/1/edit
-      page.must_have_css "form #thing_name"
-      page.must_have_css "form #thing_name.readonly"
-
-      fill_in "Description", with: "Great band"
-      click_button "Update Thing"
-
-      page.current_path.must_equal thing_path(Thing.last)
-      page.must_have_content "Great band"
-
-
-      # Remove signed in author.
-      click_link "Edit"
-
-      # check "I'm the author!"
-      check("Remove")
-      click_button "Update Thing"
-
-      page.current_path.must_equal thing_path(Thing.last)
-      page.wont_have_content "By fred@trb.org"
-    end
-
-    # this is tested in ThingCrud.
-    # # signed in: create thing without adding author.
-    # it "xxx" do
-    #   sign_in!
-
-    #   visit "/things/new"
-
-    #   fill_in 'Name', with: "Bad Religion"
-    #   click_button "Create Thing"
-
-    #   # /things/1
-    #   page.current_path.must_equal thing_path(Thing.last)
-    #   page.wont_have_content "By fred@trb.org"
-    # end
-  end
-
-
   # controller tests are what i normally do manually, per "concept"
   # see above: workflow for login, create thing, update thing, check markup, that's it.
 
-
-  describe "#create" do
-    it do
-      visit "/things/new"
-      fill_in 'Name', with: "Bad Religion"
-      click_button "Create Thing"
-
-      page.current_path.must_equal thing_path(Thing.last)
-    end
-
-    # TODO: better DSL for #post, etc.
-
-    it do # invalid.
-      # post :create, {thing: {name: ""}}
-      visit "/things/new"
-      fill_in 'Name', with: ""
-      click_button "Create Thing"
-
-      page.must_have_css ".error"
-
-      # 3 author email fields
-      page.must_have_css("input.email", count: 3)
-    end
-  end
 
   describe "#edit" do
     # not signed-in.
